requirements_txt.txt
streamlit>=1.28.0
requests>=2.31.0
Pillow>=10.0.0
python-multipart>=0.0.6
watchdog>=3.0.0